requirements.txt
feedparser==6.0.11
httpx==0.28.1
python-dotenv==1.0.1
PyYAML==6.0.2
aiosqlite==0.20.0
selectolax==0.3.21
tenacity==9.1.2
python-telegram-bot==21.9
prometheus-client==0.21.1
